requirements.txt
langchain-openai
langchain_community
langchain-huggingface
pypdf
faiss-cpu
python-dotenv